tests/wpt/mozilla/tests/bluetooth/requestDevice/not-accept-all-devices-without-filter.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return promise_rejects_js(
        t, TypeError,
        window.navigator.bluetooth.requestDevice({acceptAllDevices: false}),
        'Filters member is empty');
}, 'Reject with TypeError if `filters` not present and `acceptAllDevices` is false.');
</script>
